refactor: 필터 구조 변경으로 인한 OidcUser 상속 구현
